+++ CHANGELOG.md
@@ -10,7 +10,7 @@ for further information about branching and tagging conventions.
 ## [Unreleased]
 #### Added
 - Tooltip directive
-- "compare" method to `MathUtil` class
+- "compare" & "format" methods to `MathUtil` class
 - Name attributes to 2FA form inputs
 - Emitted value normalizers for input field:
   - "normalizeRange"
@@ -20,10 +20,14 @@ for further information about branching and tagging conventions.
 - Now loading account converted balances in
   "LOAD_ACCOUNT_BALANCES_DETAILS" vuex action
 - Now using `MathUtil.compare` method in validators
-- Now using custom number formatter if numeral returns "NaN"
+- Now using `MathUtil.format` for formatting money, numbers, and percents
 
 #### Fixed
 - Translations on sale details drawer
+
+#### Removed
+- "numeral" package
+- Order number format & filter
 
 ## [1.7.0-rc.2] - 2019-05-21
 ### "Under the hood" changes